requirements.txt
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.0
tensorflow==2.13.1
keras==2.13.1
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0
pillow==10.0.1
opencv-python==4.8.0.76


